requirements.txt
flare-capa==7.4.0
pandas==2.0.3
matplotlib==3.7.5
seaborn==0.13.2
scikit-learn==1.3.2